requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
instagrapi==2.1.1
numpy==1.26.4
panda==0.3.1
pandas==2.2.2
pillow==10.3.0
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
